sync: update from monorepo @ ae32fde946 (2026-07-28)

+++ crates/apps/symtropy-exoskeleton-demo/src/plugin.rs
@@ -97,7 +97,7 @@ fn step_exoskeleton(time: Res<Time>, mut exo: ResMut<ExoskeletonResources>) {
     let current_hv = exo.encoder.encode(&new_state);
     let pe = if let Some(ref prev) = exo.last_perception {
         let sim = current_hv.similarity(prev);
-        (1.0 - sim.max(0.0)).min(1.0) as f32
+        (1.0 - sim.max(0.0)).min(1.0)
     } else {
         0.0
     };
