+++ crates/apps/symtropy-flight-demo/src/plugin.rs
@@ -152,7 +152,7 @@ fn step_quadrotor(
     let current_hv = quad.encoder.encode(&new_state);
     let pe = if let Some(ref prev) = quad.last_perception {
         let sim = current_hv.similarity(prev);
-        (1.0 - sim.max(0.0)).min(1.0) as f32
+        (1.0 - sim.max(0.0)).min(1.0)
     } else {
         0.0
     };
