+++ crates/apps/symtropy-quadruped-demo/src/plugin.rs
@@ -100,7 +100,7 @@ fn step_quadruped(
     let current_hv = q.encoder.encode(&new_state);
     let pe = if let Some(ref prev) = q.last_perception {
         let sim = current_hv.similarity(prev);
-        (1.0 - sim.max(0.0)).min(1.0) as f32
+        (1.0 - sim.max(0.0)).min(1.0)
     } else {
         0.0
     };
