+++ crates/apps/symtropy-surgical-demo/src/plugin.rs
@@ -95,7 +95,7 @@ fn step_surgical(time: Res<Time>, mut surg: ResMut<SurgicalResources>) {
     let current_hv = surg.encoder.encode(&new_state);
     let pe = if let Some(ref prev) = surg.last_perception {
         let sim = current_hv.similarity(prev);
-        (1.0 - sim.max(0.0)).min(1.0) as f32
+        (1.0 - sim.max(0.0)).min(1.0)
     } else {
         0.0
     };
